Split graphlike search into multiple files
