+++ src/stim/search/graphlike/algo.perf.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "stim/search/graphlike/min_distance.h"
+#include "stim/search/search.h"
 
 #include "stim/benchmark_util.perf.h"
 #include "stim/gen/gen_surface_code.h"
