+++ src/stim/search/graphlike/edge.h
@@ -14,9 +14,28 @@
  * limitations under the License.
  */
 
-#ifndef _STIM_SEARCH_SEARCH_H
-#define _STIM_SEARCH_SEARCH_H
+#ifndef _STIM_SEARCH_GRAPHLIKE_EDGE_H
+#define _STIM_SEARCH_GRAPHLIKE_EDGE_H
 
-#include "stim/search/graphlike/algo.h"
+#include <cstdint>
+#include <iostream>
+#include <string>
+
+namespace stim {
+
+namespace impl_search_graphlike {
+
+/// A graphlike edge from a detector error model.
+struct Edge {
+    uint64_t opposite_node_index;
+    uint64_t crossing_observable_mask;
+    std::string str() const;
+    bool operator==(const Edge &other) const;
+    bool operator!=(const Edge &other) const;
+};
+std::ostream &operator<<(std::ostream &out, const Edge &v);
+
+}  // namespace impl_search_graphlike
+}  // namespace stim
 
 #endif
